Rewrite bot in GO
As I sat down to think about how to deploy the bot to a server, I
started to remember the painpoints with dealing with nodeJS. The server
must have nodeJS already installed, dependencies need to be downloaded,
and there has to be a way to start/restart the services when the server
goes down and comes back up.

Switching to Go makes this a bit easier. Go binaries, especially those
not compiled with CGO can be natively compiled and deployed without
worring about system dependencies. We just need to ship a binary. We'll
do this via fpm, and then deploying becomes as simple as downloading the
package, and executing it.