Attempt Meetup API integration and post _something_ to Discord
While expiramenting with Meetup API, we cannot successfully get a
authorization token yet. A email was sent to Meetup to help diagnose the
issue on their end. In the meantime, setup some testing infrastructure
to continue debugging.

Since we can't get Meetup events just yet, post _something_ to Discord
so that we have some automatic reminder.